+++ CHANGELOG.md
@@ -17,6 +17,8 @@ All notable changes to this project will be documented in this file. This projec
 ### @comet/cms-admin
 
 -   Breaking: Remove dependency @comet/react-app-auth and all occurences
+-   Breaking: Rename Publisher to PublisherPage
+-   Use data grid instead of table for PublisherPage
 -   Breaking: changed CmsBlockContext.damConfig.maxFileSize/maxSrcResolution to number (was string previously)
 
 #### Changes
@@ -30,6 +32,7 @@ All notable changes to this project will be documented in this file. This projec
 -   Restrict access to builds based on `ContentScopeModule`
 -   Decouple KubernetesModule from BuildsModule (Breaking: BuildsModuleConfig is now KubernetesModuleConfig)
 -   Breaking: AuthModule removes support for Bearer-Token
+-   Breaking: Remove export for BuildObject (should not be needed in application land)
 -   Breaking: changed DamModule damConfig.allowedImageSizes/allowedAspectRatios to number[]/string[] (was string previously)
 
 #### Changes